fixed bugs

+++ DataBaseApp/Program.cs
@@ -1,14 +1,16 @@
 ﻿using DataBaseApp.Connections;
 using DateBaseSQL.Methods;
-using Microsoft.AspNetCore.DataProtection.KeyManagement;
 using Newtonsoft.Json;
 
 public static class Program
 {
     static void Main(string[] args)
     {
-
-        File.WriteAllText("appsettings.json", "{\n  \"ConnectionString\": {\n    \"PgConnection\": \"\"\n  }\n}");
+        // Create appsettings.json if it doesn't exist
+        if (!File.Exists("appsettings.json"))
+        {
+            File.WriteAllText("appsettings.json", "{\n  \"ConnectionString\": {\n    \"PgConnection\": \"\"\n  }\n}");
+        }
 
         Console.ForegroundColor = ConsoleColor.White;
         Console.Write("Host (localhost): ");
@@ -17,15 +19,13 @@ public static class Program
         Console.Write("Port (5432): ");
         string port = Console.ReadLine();
 
-        Console.Write("Database nomi (Name): ");
+        Console.Write("Database name (Name): ");
         string database = Console.ReadLine();
 
         Console.Write("User ID (postgres): ");
         string userId = Console.ReadLine();
 
         Console.Write("Password: ");
-        //string password = Console.ReadLine();
-
         string password = string.Empty;
         ConsoleKeyInfo keyInfo;
         do
@@ -42,10 +42,7 @@ public static class Program
                 password = password.Substring(0, password.Length - 1);
                 Console.Write("\b \b");
             }
-        }
-
-        while (keyInfo.Key != ConsoleKey.Enter);
-
+        } while (keyInfo.Key != ConsoleKey.Enter);
 
         Console.Clear();
 
@@ -65,24 +62,24 @@ public static class Program
         var loadedSettings = JsonConvert.DeserializeObject<dynamic>(File.ReadAllText("appsettings.json"));
         string pgConnection = loadedSettings.ConnectionString.PgConnection;
 
-
         bool exit = false;
         int selectedIndex = 0;
-        Console.WriteLine("          Sechmas");
-        List<string> Buyruqlar = new List<string>
+        Console.WriteLine("          Menu");
+        List<string> commands = new List<string>
         {
             "Create Table",
             "List Table",
             "Alter Table",
             "Delete Table",
             "Working with Elements",
-            "Insert into Column"
+            "Insert into Column",
+            "Exit" 
         };
 
         while (!exit)
         {
             Console.Clear();
-            for (int i = 0; i < Buyruqlar.Count; i++)
+            for (int i = 0; i < commands.Count; i++)
             {
                 if (i == selectedIndex)
                 {
@@ -90,7 +87,7 @@ public static class Program
                     Console.ForegroundColor = ConsoleColor.Black;
                 }
 
-                Console.WriteLine(Buyruqlar[i]);
+                Console.WriteLine(commands[i]);
                 Console.ResetColor();
             }
 
@@ -98,42 +95,43 @@ public static class Program
 
             if (key.Key == ConsoleKey.DownArrow)
             {
-                selectedIndex = (selectedIndex + 1) % Buyruqlar.Count;
+                selectedIndex = (selectedIndex + 1) % commands.Count;
             }
             else if (key.Key == ConsoleKey.UpArrow)
             {
-                selectedIndex = (selectedIndex - 1 + Buyruqlar.Count) % Buyruqlar.Count;
+                selectedIndex = (selectedIndex - 1 + commands.Count) % commands.Count;
             }
             else if (key.Key == ConsoleKey.Enter)
             {
                 switch (selectedIndex)
                 {
                     case 0:
-                        Create.CreateTable(connectionString);
+                        Create.CreateTable(pgConnection);
                         break;
                     case 1:
-                        Select.GetTableNames(connectionString);
+                        Select.GetTableNames(pgConnection);
                         break;
                     case 2:
-                        Alter.AlterTable(connectionString);
+                        Alter.AlterTable(pgConnection);
                         break;
                     case 3:
-                        Deleting.DeleteTable(connectionString);
+                        Deleting.DeleteTable(pgConnection);
                         break;
                     case 4:
-                        Elements.Elementss(connectionString);
+                        Elements.Elementss(pgConnection);
                         break;
-                    case 5: 
-                        Insert.InsertIntoColumn(connectionString);
+                    case 5:
+                        Insert.InsertIntoColumn(pgConnection);
+                        break;
+                    case 6:
+                        exit = true;
                         break;
                     default:
                         Console.WriteLine("Error!!!");
                         break;
-
                 }
 
-                Console.ReadKey();
-
+                if (!exit) Console.ReadKey();
             }
         }
         Console.Clear();
